// packages/papaya/frontend/sirius-components-omnibox/src/Omnibox.types.ts
/*******************************************************************************
 * Copyright (c) 2024 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/

export interface OmniboxProps {
  initialContextEntries: OmniboxContextEntry[];
  open: boolean;
  onClose: () => void;
}

export interface OmniboxState {
  contextEntries: OmniboxContextEntry[];
  query: string;
  actions: OmniboxAction[];
}

export interface OmniboxContextEntry {
  id: string;
  label: string;
  kind: OmniboxContextKind;
}

export type OmniboxContextKind = 'EditingContext';

export interface OmniboxAction {
  id: string;
  icon: JSX.Element;
  label: string;
}